PQVPN/requirements.txt
PySide6>=6.5
PyYAML>=6.0
cryptography>=41.0.0
asyncudp>=0.1.0
argon2-cffi>=21.3.0
liboqs-python
psutil>=5.9.0
dnspython>=2.3.0
requests>=2.28.0
pyOpenSSL>=23.0.0
pycryptodome>=3.17.0
cffi>=1.15.0
pycparser>=2.21
# Windows TUN support (install one of these on Windows)
wintun>=0.1.0; platform_system == "Windows"
python-pytun>=2.3.0; platform_system == "Windows"

